Adding the custom version of Erela.js!
I have implemented a custom version of Erela.js created by Tomato6966. It supports both the REST API and WebSocket, even for users using an older version of Lavalink. The custom version of Erela.js supports both v3 and v4 of the REST API, allowing users to choose which version they prefer, and even v2 if necessary.

Special thanks to Tomato6966 for their updates and improvements.